Rename code_style dir to style
